--- a/Explorer/Assets/DCL/Infrastructure/ECS/Arch.System.SourceGenerator.dll
+++ b/Explorer/Assets/DCL/Infrastructure/ECS/Arch.System.SourceGenerator.dll
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:308914dfaad9fc92a85c16e606153a70ba5027dddb100e43847a35c5f15fbd59
-size 27648
+oid sha256:aa3a0748fb8627d5295feacdb2ed40ec2bf6ad2c3b8e7f43b5ab6a4eabe02b4c
+size 32768