--- a/Explorer/Assets/DCL/Interaction/Systems/ProcessOtherAvatarsInteractionSystem.cs
+++ b/Explorer/Assets/DCL/Interaction/Systems/ProcessOtherAvatarsInteractionSystem.cs
@@ -65,10 +65,10 @@ namespace DCL.Interaction.Systems
             if (!raycastResultForGlobalEntities.IsValidHit || !canHover || entityInfo == null)
                 return;
 
-            EntityReference entityRef = entityInfo.Value.EntityReference;
+            Entity entityRef = entityInfo.Value.EntityReference;
 
-            if (!entityRef.IsAlive(World!) || !World!.TryGet(entityRef, out Profile? profile)
-                                           || World.Has<BlockedPlayerComponent>(entityRef))
+            if (!World.IsAlive(entityRef) || !World!.TryGet(entityRef, out Profile? profile)
+                                          || World.Has<BlockedPlayerComponent>(entityRef))
                 return;
 
             currentProfileHovered = profile;