… Explorer/Assets/DCL/AvatarRendering/AvatarShape/Tests/EditMode/AvatarLoaderSystemShould.cs
… Explorer/Assets/DCL/AvatarRendering/Wearables/Tests/ResolveWearableByPointerSystemShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/LifeCycle/Tests/ReleaseRemovedComponentsSystemShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/SceneLifeCycle/Tests/LoadFixedPointersShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/SceneLifeCycle/Tests/ResolveSceneStateByIncreasingRadiusSystemShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/SceneLifeCycle/Tests/UnloadSceneSystemShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/StreamableLoading/Tests/AssetPromiseShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/Unity/GLTFContainer/Tests/CleanUpGltfContainerSystemShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/Unity/GLTFContainer/Tests/FinalizeGltfContainerLoadingSystemShould.cs
… Explorer/Assets/DCL/Infrastructure/ECS/Unity/Transforms/Systems/ParentingTransformSystem.cs
… Explorer/Assets/DCL/Infrastructure/ECS/Unity/Transforms/Tests/ParentingTransformSystemShould.cs
… Explorer/Assets/DCL/Interaction/PlayerOriginated/Tests/ProcessPointerEventsSystemShould.cs
… Explorer/Assets/DCL/Interaction/Raycast/Components/HighlightComponent.cs
… Explorer/Assets/DCL/Interaction/Raycast/Tests/ExecuteRaycastSystemShould.cs
… Explorer/Assets/DCL/Interaction/Systems/InteractionHighlightSystem.cs
… Explorer/Assets/DCL/SDKComponents/Animator/Tests/AnimatorHandlerShould.cs
… Explorer/Assets/DCL/SDKComponents/AvatarAttach/Systems/InstantiateTransformForAvatarAttachSystem.cs
… Explorer/Assets/DCL/SDKComponents/AvatarModifierArea/Systems/AvatarModifierAreaHandlerSystem.cs
… Explorer/Assets/DCL/SDKComponents/NFTShape/Tests/CleanUpNftShapeSystemShould.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Components/UITransformComponent.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Components/UITransformRelationLinkedData.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Systems/UITransform/UITransformParentingSystem.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Systems/UITransform/UITransformSortingSystem.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Tests/UITransformInstantiationSystemShould.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Tests/UITransformParentingSystemShould.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Tests/UITransformSortingSystemShould.cs
… Explorer/Assets/DCL/SDKComponents/SceneUI/Tests/UITransformUpdateSystemShould.cs
… Explorer/Assets/DCL/SDKComponents/Tween/Systems/TweenLoaderSystem.cs
… Explorer/Assets/Plugins/Arch/Arch.LowLevel.dll
… Explorer/Assets/Plugins/Arch/Arch.LowLevel.xml