## Benchmarks after:

|            Method |  Amount |         Mean |       Error |      StdDev |       Median | Allocated |
|------------------ |-------- |-------------:|------------:|------------:|-------------:|----------:|
|  WorldEntityQuery |   10000 |    55.619 us |   0.6289 us |   0.5883 us |    55.605 us |         - |
|             Query |   10000 |     7.304 us |   0.0788 us |   0.0658 us |     7.301 us |         - |
|       EntityQuery |   10000 |     9.238 us |   0.1196 us |   0.1119 us |     9.244 us |         - |
|          Iterator |   10000 |     6.663 us |   0.1188 us |   0.1111 us |     6.657 us |         - |
|       StructQuery |   10000 |     6.558 us |   0.1098 us |   0.1027 us |     6.567 us |         - |
| StructEntityQuery |   10000 |     7.298 us |   0.0447 us |   0.0374 us |     7.309 us |         - |
|  WorldEntityQuery |  100000 |   557.017 us |   4.0447 us |   3.7834 us |   556.285 us |       1 B |
|             Query |  100000 |    74.572 us |   0.8722 us |   0.8158 us |    74.550 us |         - |
|       EntityQuery |  100000 |    95.645 us |   1.9127 us |   1.6955 us |    95.401 us |         - |
|          Iterator |  100000 |    67.907 us |   1.2780 us |   1.1955 us |    67.492 us |         - |
|       StructQuery |  100000 |    69.120 us |   1.2344 us |   1.0943 us |    68.788 us |         - |
| StructEntityQuery |  100000 |    78.074 us |   0.9875 us |   0.9237 us |    77.911 us |         - |
|  WorldEntityQuery | 1000000 | 6,186.398 us | 121.8281 us | 287.1632 us | 6,101.882 us |       5 B |
|             Query | 1000000 |   856.740 us |   9.5109 us |   8.4312 us |   857.373 us |       1 B |
|       EntityQuery | 1000000 | 1,330.767 us |  30.0280 us |  88.5383 us | 1,320.457 us |       1 B |
|          Iterator | 1000000 |   791.606 us |  14.8775 us |  34.1835 us |   777.363 us |       1 B |
|       StructQuery | 1000000 |   760.958 us |   8.1763 us |   7.6481 us |   761.836 us |       1 B |
| StructEntityQuery | 1000000 | 1,016.722 us |  17.2361 us |  14.3929 us | 1,014.765 us |       1 B |

|                                Method |  Amount |        Mean |     Error |    StdDev |   Gen0 | Allocated |
|-------------------------------------- |-------- |------------:|----------:|----------:|-------:|----------:|
|          IterationNormalTwoComponents |   10000 |    18.06 us |  0.279 us |  0.233 us |      - |         - |
|       IterationUnsafeAddTwoComponents |   10000 |    18.22 us |  0.360 us |  0.337 us |      - |         - |
|            IterationParallelUnsafeAdd |   10000 |    12.55 us |  0.229 us |  0.214 us | 0.3357 |    5388 B |
|    IterationNormalEntityTwoComponents |   10000 |    27.27 us |  0.398 us |  0.352 us |      - |         - |
| IterationUnsafeAddEntityTwoComponents |   10000 |    25.77 us |  0.482 us |  0.495 us |      - |         - |
|          IterationNormalTwoComponents |  100000 |   177.44 us |  3.395 us |  3.176 us |      - |         - |
|       IterationUnsafeAddTwoComponents |  100000 |   184.91 us |  3.381 us |  3.163 us |      - |         - |
|            IterationParallelUnsafeAdd |  100000 |    57.62 us |  1.151 us |  1.182 us | 0.4272 |    7037 B |
|    IterationNormalEntityTwoComponents |  100000 |   271.95 us |  5.024 us |  4.699 us |      - |         - |
| IterationUnsafeAddEntityTwoComponents |  100000 |   270.27 us |  5.402 us |  5.306 us |      - |         - |
|          IterationNormalTwoComponents | 1000000 | 1,816.22 us | 15.214 us | 14.231 us |      - |         - |
|       IterationUnsafeAddTwoComponents | 1000000 | 1,640.38 us | 31.814 us | 35.361 us |      - |         - |
|            IterationParallelUnsafeAdd | 1000000 |   345.59 us |  6.897 us | 14.993 us | 0.4883 |   10823 B |
|    IterationNormalEntityTwoComponents | 1000000 | 2,806.90 us | 32.268 us | 28.605 us |      - |         - |
| IterationUnsafeAddEntityTwoComponents | 1000000 | 2,710.09 us | 30.188 us | 28.237 us |      - |         - |



## Benchmarks before:


|            Method |  Amount |          Mean |       Error |     StdDev | Allocated |
|------------------ |-------- |--------------:|------------:|-----------:|----------:|
|  WorldEntityQuery |   10000 |    165.357 us |   2.3075 us |  1.9269 us |         - |
|             Query |   10000 |     16.665 us |   0.3320 us |  0.4433 us |         - |
|       EntityQuery |   10000 |     40.473 us |   0.1099 us |  0.1028 us |         - |
|          Iterator |   10000 |      6.510 us |   0.0893 us |  0.0836 us |         - |
|       StructQuery |   10000 |      6.119 us |   0.0182 us |  0.0142 us |         - |
| StructEntityQuery |   10000 |      7.897 us |   0.0271 us |  0.0212 us |         - |
|  WorldEntityQuery |  100000 |  1,654.528 us |  17.5350 us | 16.4023 us |       2 B |
|             Query |  100000 |    145.848 us |   1.1623 us |  1.0872 us |         - |
|       EntityQuery |  100000 |    401.757 us |   1.2116 us |  1.1334 us |         - |
|          Iterator |  100000 |     64.183 us |   0.3787 us |  0.3162 us |         - |
|       StructQuery |  100000 |     63.844 us |   0.9119 us |  0.8084 us |         - |
| StructEntityQuery |  100000 |     80.777 us |   1.0128 us |  0.9474 us |         - |
|  WorldEntityQuery | 1000000 | 16,466.491 us | 101.2828 us | 89.7845 us |      26 B |
|             Query | 1000000 |  1,512.173 us |  12.1726 us | 10.7907 us |       2 B |
|       EntityQuery | 1000000 |  4,040.493 us |   9.2860 us |  7.7542 us |       6 B |
|          Iterator | 1000000 |    789.160 us |  11.2362 us | 10.5103 us |       1 B |
|       StructQuery | 1000000 |    799.028 us |  15.6162 us | 19.7495 us |       1 B |
| StructEntityQuery | 1000000 |  1,084.173 us |  21.1478 us | 26.7452 us |       2 B |

|                                Method |  Amount |         Mean |      Error |     StdDev |   Gen0 | Allocated |
|-------------------------------------- |-------- |-------------:|-----------:|-----------:|-------:|----------:|
|          IterationNormalTwoComponents |   10000 |    19.209 us |  0.3794 us |  0.4060 us |      - |         - |
|       IterationUnsafeAddTwoComponents |   10000 |    16.850 us |  0.1150 us |  0.1076 us |      - |         - |
|            IterationParallelUnsafeAdd |   10000 |     9.592 us |  0.1519 us |  0.1421 us | 0.3204 |    5181 B |
|    IterationNormalEntityTwoComponents |   10000 |    28.121 us |  0.4768 us |  0.4460 us |      - |         - |
| IterationUnsafeAddEntityTwoComponents |   10000 |    26.272 us |  0.1947 us |  0.1822 us |      - |         - |
|          IterationNormalTwoComponents |  100000 |   187.334 us |  1.1285 us |  1.0004 us |      - |         - |
|       IterationUnsafeAddTwoComponents |  100000 |   164.108 us |  1.3256 us |  1.2400 us |      - |         - |
|            IterationParallelUnsafeAdd |  100000 |    47.987 us |  0.5137 us |  0.4805 us | 0.4883 |    8129 B |
|    IterationNormalEntityTwoComponents |  100000 |   265.345 us |  5.2661 us |  7.5524 us |      - |         - |
| IterationUnsafeAddEntityTwoComponents |  100000 |   279.013 us |  5.2673 us |  4.9270 us |      - |         - |
|          IterationNormalTwoComponents | 1000000 | 1,880.585 us | 12.9674 us | 12.1297 us |      - |       2 B |
|       IterationUnsafeAddTwoComponents | 1000000 | 1,653.831 us | 27.0212 us | 25.2756 us |      - |       2 B |
|            IterationParallelUnsafeAdd | 1000000 |   343.740 us |  7.3164 us | 21.3422 us | 0.4883 |   10334 B |
|    IterationNormalEntityTwoComponents | 1000000 | 2,847.196 us | 40.0667 us | 37.4784 us |      - |       3 B |
| IterationUnsafeAddEntityTwoComponents | 1000000 | 2,841.296 us | 48.9425 us | 45.7808 us |      - |       3 B |

## Results:

| Method            | Amount  | Mean Before (us) | Mean After (us) | Difference (us) | % Change |
|------------------|---------|------------------|------------------|------------------|----------|
| WorldEntityQuery | 10000   | 165.357          | 55.619           | 109.738          | 66.36%   |
| Query            | 10000   | 16.665           | 7.304            | 9.361            | 56.17%   |
| EntityQuery      | 10000   | 40.473           | 9.238            | 31.235           | 77.17%   |
| Iterator         | 10000   | 6.510            | 6.663            | -0.153           | -2.35%   |
| StructQuery      | 10000   | 6.119            | 6.558            | -0.439           | -7.17%   |
| StructEntityQuery| 10000   | 7.897            | 7.298            | 0.599            | 7.59%    |
| WorldEntityQuery | 100000  | 1654.528         | 557.017          | 1097.511         | 66.34%   |
| Query            | 100000  | 145.848          | 74.572           | 71.276           | 48.88%   |
| EntityQuery      | 100000  | 401.757          | 95.645           | 306.112          | 76.21%   |
| Iterator         | 100000  | 64.183           | 67.907           | -3.724           | -5.80%   |
| StructQuery      | 100000  | 63.844           | 69.120           | -5.276           | -8.26%   |
| StructEntityQuery| 100000  | 80.777           | 78.074           | 2.703            | 3.35%    |
| WorldEntityQuery | 1000000 | 16466.491        | 6186.398         | 10280.093        | 62.43%   |
| Query            | 1000000 | 1512.173         | 856.740          | 655.433          | 43.34%   |
| EntityQuery      | 1000000 | 4040.493         | 1330.767         | 2709.726         | 67.05%   |
| Iterator         | 1000000 | 789.160          | 791.606          | -2.446           | -0.31%   |
| StructQuery      | 1000000 | 799.028          | 760.958          | 38.070           | 4.76%    |
| StructEntityQuery| 1000000 | 1084.173         | 1016.722         | 67.451           | 6.22%    |

| Method                               | Amount  | Mean Before (us) | Mean After (us) | Difference (us) | % Change |
|-------------------------------------|---------|------------------|------------------|------------------|----------|
| IterationNormalTwoComponents        | 10000   | 19.209           | 18.060           | 1.149            | 5.98%    |
| IterationUnsafeAddTwoComponents     | 10000   | 16.850           | 18.220           | -1.370           | -8.13%   |
| IterationParallelUnsafeAdd          | 10000   | 9.592            | 12.550           | -2.958           | -30.84%  |
| IterationNormalEntityTwoComponents  | 10000   | 28.121           | 27.270           | 0.851            | 3.03%    |
| IterationUnsafeAddEntityTwoComponents| 10000  | 26.272           | 25.770           | 0.502            | 1.91%    |
| IterationNormalTwoComponents        | 100000  | 187.334          | 177.440          | 9.894            | 5.28%    |
| IterationUnsafeAddTwoComponents     | 100000  | 164.108          | 184.910          | -20.802          | -12.68%  |
| IterationParallelUnsafeAdd          | 100000  | 47.987           | 57.620           | -9.633           | -20.07%  |
| IterationNormalEntityTwoComponents  | 100000  | 265.345          | 271.950          | -6.605           | -2.49%   |
| IterationUnsafeAddEntityTwoComponents| 100000 | 279.013          | 270.270          | 8.743            | 3.13%    |
| IterationNormalTwoComponents        | 1000000 | 1880.585         | 1816.220         | 64.365           | 3.42%    |
| IterationUnsafeAddTwoComponents     | 1000000 | 1653.831         | 1640.380         | 13.451           | 0.81%    |
| IterationParallelUnsafeAdd          | 1000000 | 343.740          | 345.590          | -1.850           | -0.54%   |
| IterationNormalEntityTwoComponents  | 1000000 | 2847.196         | 2806.900         | 40.296           | 1.41%    |
| IterationUnsafeAddEntityTwoComponents|1000000 | 2841.296         | 2710.090         | 131.206          | 4.62%    |

## Conclusions:

There are major improvements in all kinds of queries, iterations themselves are not affected in general: `Parallel` performance we can ignore as we don't use it at all


## How To QA

Perform Regression